Loading during native splash screen
This commit moves the app settings loading from the splash screen page into the native splash screen.
